# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface Profile {
  id: string;
  email: string;
  full_name: string;
  username: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  wallet_balance: number;
  is_pin_set: boolean;
  created_at?: string;
}

export interface PinInput {
  /**
     * @minLength 4
     * @maxLength 4
     */
  pin: string;
}

export interface PinVerifyResponse {
  valid: boolean;
}

export interface DataPlan {
  id: number;
  network: string;
  plan_name: string;
  data_size: string;
  retail_price: number;
  wholesale_price?: number;
  plan_id: string;
  /** @nullable */
  validity?: string | null;
}

export interface DataPurchaseInput {
  phone: string;
  plan_id: string;
  network?: string;
  pin: string;
}

export interface AirtimePurchaseInput {
  phone: string;
  network: string;
  amount: number;
  pin: string;
}

export interface CablePurchaseInput {
  smart_card_number: string;
  cable_provider: string;
  plan_id: string;
  pin: string;
}

export interface ElectricityPurchaseInput {
  meter_number: string;
  disco: string;
  amount: number;
  meter_type: string;
  pin: string;
}

export interface ServiceResponse {
  success: boolean;
  message: string;
  /** @nullable */
  token?: string | null;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  new_balance?: number | null;
}

export interface WalletFundInput {
  amount: number;
}

export interface WalletFundResponse {
  base_amount: number;
  processing_fee: number;
  total_amount: number;
  reference: string;
  payment_instruction?: string;
}

export interface WalletTransaction {
  id: string;
  type: string;
  amount: number;
  description: string;
  status: string;
  /** @nullable */
  reference?: string | null;
  created_at: string;
}

export interface WalletSummary {
  balance: number;
  total_funded: number;
  total_spent: number;
  transaction_count: number;
  data_purchases?: number;
  airtime_purchases?: number;
}

export interface AdminUser {
  id: string;
  email: string;
  full_name: string;
  username: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  wallet_balance: number;
  is_pin_set?: boolean;
  created_at: string;
}

export interface AdminTransaction {
  id: string;
  user_id: string;
  user_email: string;
  user_name?: string;
  type: string;
  amount: number;
  description: string;
  status: string;
  /** @nullable */
  reference?: string | null;
  created_at: string;
}

export interface AdminStats {
  total_users: number;
  total_transactions: number;
  total_revenue: number;
  total_disbursed: number;
  today_transactions?: number;
  today_revenue?: number;
  active_users_today?: number;
}

export interface SystemSettings {
  id: number;
  active_payment_gateway: string;
  /** @nullable */
  cheapdatahub_funding_account?: string | null;
  /** @nullable */
  updated_at?: string | null;
}

export interface SystemSettingsUpdate {
  active_payment_gateway?: string;
  /** @nullable */
  paystack_secret_key?: string | null;
  /** @nullable */
  flutterwave_secret_key?: string | null;
  /** @nullable */
  cheapdatahub_api_key?: string | null;
  /** @nullable */
  cheapdatahub_funding_account?: string | null;
  /** @nullable */
  brevo_api_key?: string | null;
}

export type WebhookPayloadData = { [key: string]: unknown };

export interface WebhookPayload {
  event?: string;
  data?: WebhookPayloadData;
  [key: string]: unknown;
 }

export type AdminGetTransactionsParams = {
/**
 * Filter by date (YYYY-MM-DD)
 * @nullable
 */
date?: string | null;
};

export type AdminExportTransactionsParams = {
/**
 * @nullable
 */
date?: string | null;
};