# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CheapDataHub VTU & Bill Payment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/pin:
    post:
      operationId: setupPin
      tags: [profile]
      summary: Set transaction PIN (first-time setup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "200":
          description: PIN set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /profile/pin/verify:
    post:
      operationId: verifyPin
      tags: [profile]
      summary: Verify transaction PIN before a service purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "200":
          description: PIN verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinVerifyResponse"

  # ─── Data Plans ─────────────────────────────────────────────────────────────
...
  /data-plans/{network}:
    get:
      operationId: getDataPlansByNetwork
      tags: [data-plans]
      summary: Get data plans for a specific network
      parameters:
        - name: network
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Network data plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataPlan"

  # ─── Services ───────────────────────────────────────────────────────────────
...
  /admin/transactions:
    get:
      operationId: adminGetTransactions
      tags: [admin]
      summary: Get all transactions across all users (admin only)
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
          description: Filter by date (YYYY-MM-DD)
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTransaction"
  /admin/transactions/export:
    get:
      operationId: adminExportTransactions
      tags: [admin]
      summary: Export all transactions as CSV download
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get system settings
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    PinInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          minLength: 4
          maxLength: 4
    PinVerifyResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
    DataPlan:
      type: object
      required: [id, network, plan_name, data_size, retail_price, plan_id]
      properties:
        id:
          type: integer
        network:
          type: string
        plan_name:
          type: string
        data_size:
          type: string
        retail_price:
          type: number
        wholesale_price:
          type: number
        plan_id:
          type: string
        validity:
          type: ["string", "null"]
...
    AdminTransaction:
      type: object
      required: [id, user_id, user_email, type, amount, description, status, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        user_email:
          type: string
        user_name:
          type: string
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        status:
          type: string
        reference:
          type: ["string", "null"]
        created_at:
          type: string
...
    SystemSettings:
      type: object
      required: [id, active_payment_gateway]
      properties:
        id:
          type: integer
        active_payment_gateway:
          type: string
        cheapdatahub_funding_account:
          type: ["string", "null"]
        updated_at:
          type: ["string", "null"]
    SystemSettingsUpdate:
      type: object
      properties:
        active_payment_gateway:
          type: string
        paystack_secret_key:
          type: ["string", "null"]
        flutterwave_secret_key:
          type: ["string", "null"]
        cheapdatahub_api_key:
          type: ["string", "null"]
        cheapdatahub_funding_account:
          type: ["string", "null"]
        brevo_api_key:
          type: ["string", "null"]
...