# lib/api-zod/src/generated/types/dataPlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface DataPlan {
  id: number;
  network: string;
  plan_name: string;
  data_size: string;
  retail_price: number;
  wholesale_price?: number;
  plan_id: string;
  /** @nullable */
  validity?: string | null;
}