# lib/api-zod/src/generated/types/pinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface PinInput {
  /**
     * @minLength 4
     * @maxLength 4
     */
  pin: string;
}