# lib/api-zod/src/generated/types/systemSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemSettings {
  id: number;
  active_payment_gateway: string;
  /** @nullable */
  cheapdatahub_funding_account?: string | null;
  /** @nullable */
  updated_at?: string | null;
}