# lib/api-zod/src/generated/types/systemSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CheapDataHub VTU & Bill Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemSettingsUpdate {
  active_payment_gateway?: string;
  /** @nullable */
  paystack_secret_key?: string | null;
  /** @nullable */
  flutterwave_secret_key?: string | null;
  /** @nullable */
  cheapdatahub_api_key?: string | null;
  /** @nullable */
  cheapdatahub_funding_account?: string | null;
  /** @nullable */
  brevo_api_key?: string | null;
}